+++ hutool-core/src/main/java/cn/hutool/core/img/Img.java
@@ -3,6 +3,7 @@ package cn.hutool.core.img;
 import java.awt.AlphaComposite;
 import java.awt.Color;
 import java.awt.Font;
+import java.awt.FontMetrics;
 import java.awt.Graphics2D;
 import java.awt.Image;
 import java.awt.Rectangle;
@@ -258,8 +259,10 @@ public class Img {
 		// 透明度
 		g.setComposite(AlphaComposite.getInstance(AlphaComposite.SRC_ATOP, alpha));
 		// 在指定坐标绘制水印文字
-		final int fontSize = font.getSize();
-		g.drawString(pressText, (destImage.getWidth() - (getLength(pressText) * fontSize)) / 2 + x, (destImage.getHeight() - fontSize) / 2 + y);
+		final FontMetrics metrics = g.getFontMetrics(font);
+		final int textLength = metrics.stringWidth(pressText);
+		final int textHeight = metrics.getAscent() - metrics.getLeading() - metrics.getDescent();
+		g.drawString(pressText, Math.abs(destImage.getWidth() - textLength) / 2 + x, Math.abs(destImage.getHeight() + textHeight) / 2 + y);
 		g.dispose();
 		this.destImage = destImage;
 
@@ -277,16 +280,26 @@ public class Img {
 	 * @return this
 	 */
 	public Img pressImage(Image pressImg, int x, int y, float alpha) {
-		final BufferedImage destImg = this.srcImage;
-		final int width = destImg.getWidth();
-		final int height = destImg.getHeight();
-
-		int pressImgWidth = pressImg.getWidth(null);
-		int pressImgHeight = pressImg.getHeight(null);
-		x += (width - pressImgWidth) / 2;
-		y += (height - pressImgHeight) / 2;
+		final int pressImgWidth = pressImg.getWidth(null);
+		final int pressImgHeight = pressImg.getHeight(null);
 		
-		draw(destImg, pressImg, new Rectangle(x, y, pressImgWidth, pressImgHeight), alpha);
+		return pressImage(pressImg, new Rectangle(x, y, pressImgWidth, pressImgHeight), alpha);
+	}
+	
+	/**
+	 * 给图片添加图片水印<br>
+	 * 此方法并不关闭流
+	 * 
+	 * @param pressImg 水印图片，可以使用{@link ImageIO#read(File)}方法读取文件
+	 * @param rectangle 矩形对象，表示矩形区域的x，y，width，height，x,y从背景图片中心计算
+	 * @param alpha 透明度：alpha 必须是范围 [0.0, 1.0] 之内（包含边界值）的一个浮点数字
+	 * @return this
+	 * @since 4.1.14
+	 */
+	public Img pressImage(Image pressImg,Rectangle rectangle, float alpha) {
+		final BufferedImage destImg = this.srcImage;
+		
+		draw(destImg, pressImg, rectangle, alpha);
 		this.destImage = destImg;
 		return this;
 	}
@@ -390,36 +403,26 @@ public class Img {
 	 * 
 	 * @param backgroundImg 背景图片
 	 * @param img 要绘制的图片
-	 * @param rectangle 矩形对象，表示矩形区域的x，y，width，height
+	 * @param rectangle 矩形对象，表示矩形区域的x，y，width，height，x,y从背景图片中心计算
 	 * @return 绘制后的背景
 	 */
 	private static BufferedImage draw(BufferedImage backgroundImg, Image img, Rectangle rectangle, float alpha) {
 		final Graphics2D g = backgroundImg.createGraphics();
 		g.setComposite(AlphaComposite.getInstance(AlphaComposite.SRC_ATOP, alpha));
+		
+		//修正图片位置从背景的中心计算
+		final int width = backgroundImg.getWidth();
+		final int height = backgroundImg.getHeight();
+		rectangle.setLocation(//
+				rectangle.x + (int)(Math.abs(width - rectangle.width) / 2), //
+				rectangle.y + (int)(Math.abs(height - rectangle.height) / 2)//
+		);
+		
 		g.drawImage(img, rectangle.x, rectangle.y, rectangle.width, rectangle.height, null); // 绘制切割后的图
 		g.dispose();
 		return backgroundImg;
 	}
 
-	/**
-	 * 计算text的长度（一个中文算两个字符）<br>
-	 * 如：text="中国",返回 2；text="test",返回 2；text="中国ABC",返回 4.
-	 * 
-	 * @param text 文本
-	 * @return 字符长度
-	 */
-	private static int getLength(String text) {
-		int length = 0;
-		for (int i = 0; i < text.length(); i++) {
-			if (String.valueOf(text.charAt(i)).getBytes().length > 1) {
-				length += 2;
-			} else {
-				length += 1;
-			}
-		}
-		return length / 2;
-	}
-	
 	/**
 	 * 获取int类型的图片类型
 	 *