+++ hutool-core/src/main/java/cn/hutool/core/swing/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Swing和awt相关封装
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.core.swing;