+++ hutool-core/src/test/java/cn/hutool/core/swing/RobotUtilTest.java
@@ -1,0 +1,16 @@
+package cn.hutool.core.swing;
+
+import org.junit.Ignore;
+import org.junit.Test;
+
+import cn.hutool.core.io.FileUtil;
+import cn.hutool.core.swing.RobotUtil;
+
+public class RobotUtilTest {
+	
+	@Test
+	@Ignore
+	public void captureScreenTest() {
+		RobotUtil.captureScreen(FileUtil.file("e:/screen.jpg"));
+	}
+}